Data.IntMap.Internal.Debug

Documentation

showTree :: Show a => IntMap a -> String

O(n). Show the tree that implements the map. The tree is shown in a compressed, hanging format.

showTreeWith :: Show a => Bool -> Bool -> IntMap a -> String

O(n). The expression (showTreeWith hang wide map) shows the tree that implements the map. If hang is True, a hanging tree is shown otherwise a rotated tree is shown. If wide is True, an extra wide version is shown.