Data.Sequence

Description

Finite sequences

The Seq a type represents a finite sequence of values of type a.

Sequences generally behave very much like lists.

The class instances for sequences are all based very closely on those for lists.

Many functions in this module have the same names as functions in the Prelude or in Data.List. In almost all cases, these functions behave analogously. For example, filter filters a sequence in exactly the same way that Prelude.filter filters a list. The only major exception is the lookup function, which is based on the function by that name in Data.IntMap rather than the one in Prelude.

There are two major differences between sequences and lists:

Sequences support a wider variety of efficient operations than do lists. Notably, they offer

Constant-time access to both the front and the rear with <|, |>, viewl, viewr. For recent GHC versions, this can be done more conveniently using the bidirectional patterns Empty, :<|, and :|>. See the detailed explanation in the "Pattern synonyms" section.

Logarithmic-time concatenation with ><

Logarithmic-time splitting with splitAt, take and drop

Logarithmic-time access to any element with lookup, !?, index, insertAt, deleteAt, adjust', and update

Note that sequences are typically slower than lists when using only operations for which they have the same big-(O) complexity: sequences make rather mediocre stacks!

Whereas lists can be either finite or infinite, sequences are always finite. As a result, a sequence is strict in its length. Ignoring efficiency, you can imagine that Seq is defined

 data Seq a = Empty | a :<| !(Seq a)

This means that many operations on sequences are stricter than those on lists. For example,

 (1 : undefined) !! 0 = 1

but

 (1 :<| undefined) index 0 = undefined

Sequences may also be compared to immutable arrays or vectors. Like these structures, sequences support fast indexing, although not as fast. But editing an immutable array or vector, or combining it with another, generally requires copying the entire structure; sequences generally avoid that, copying only the portion that has changed.

Detailed performance information

An amortized running time is given for each operation, with n referring to the length of the sequence and i being the integral index used by some operations. These bounds hold even in a persistent (shared) setting.

Despite sequences being structurally strict from a semantic standpoint, they are in fact implemented using laziness internally. As a result, many operations can be performed incrementally, producing their results as they are demanded. This greatly improves performance in some cases. These functions include

The Functor methods fmap and <$, along with mapWithIndex

The Applicative methods <*>, *>, and <*

The zips: zipWith, zip, etc.

heads and tails

fromFunction, replicate, intersperse, and cycleTaking

reverse

chunksOf

Note that the Monad method, >>=, is not particularly lazy. It will take time proportional to the sum of the logarithms of the individual result sequences to produce anything whatsoever.

Several functions take special advantage of sharing to produce results using much less time and memory than one might expect. These are documented individually for functions, but also include the methods <$ and *>, each of which take time and space proportional to the logarithm of the size of the result.

Warning

The size of a Seq must not exceed maxBound::Int. Violation of this condition is not detected and if the size limit is exceeded, the behaviour of the sequence is undefined. This is unlikely to occur in most applications, but some care may be required when using ><, <*>, *>, or >>, particularly repeatedly and particularly in combination with replicate or fromFunction.

Implementation

The implementation uses 2-3 finger trees annotated with sizes, as described in section 4.2 of

Ralf Hinze and Ross Paterson, "Finger trees: a simple general-purpose data structure", Journal of Functional Programming 16:2 (2006) pp 197-217.

Finite sequences

data Seq a where

General-purpose finite sequences.

Bundled Patterns

pattern Empty :: Seq a
A bidirectional pattern synonym matching an empty sequence.

Since: 0.5.8

pattern (:<|) :: a -> Seq a -> Seq a infixr 5
A bidirectional pattern synonym viewing the front of a non-empty sequence.

Since: 0.5.8

pattern (:|>) :: Seq a -> a -> Seq a infixl 5
A bidirectional pattern synonym viewing the rear of a non-empty sequence.

Since: 0.5.8

Pattern synonyms

Much like lists can be constructed and matched using the : and [] constructors, sequences can be constructed and matched using the Empty, :<|, and :|> pattern synonyms.

Note

These patterns are only available with GHC version 8.0 or later, and version 8.2 works better with them. When writing for such recent versions of GHC, the patterns can be used in place of empty, <|, |>, viewl, and viewr.

Pattern synonym examples

Expand

Import the patterns:

import Data.Sequence (Seq (..))

Look at the first three elements of a sequence

getFirst3 :: Seq a -> Maybe (a,a,a)
getFirst3 (x1 :<| x2 :<| x3 :<| _xs) = Just (x1,x2,x3)
getFirst3 _ = Nothing

> getFirst3 (fromList [1,2,3,4]) = Just (1,2,3)
> getFirst3 (fromList [1,2]) = Nothing

Move the last two elements from the end of the first list onto the beginning of the second one.

shift2Right :: Seq a -> Seq a -> (Seq a, Seq a)
shift2Right Empty ys = (Empty, ys)
shift2Right (Empty :|> x) ys = (Empty, x :<| ys)
shift2Right (xs :|> x1 :|> x2) = (xs, x1 :<| x2 :<| ys)

> shift2Right (fromList []) (fromList [10]) = (fromList [], fromList [10])
> shift2Right (fromList [9]) (fromList [10]) = (fromList [], fromList [9,10])
> shift2Right (fromList [8,9]) (fromList [10]) = (fromList [], fromList [8,9,10])
> shift2Right (fromList [7,8,9]) (fromList [10]) = (fromList [7], fromList [8,9,10])

Construction

empty :: Seq a

\( O(1) \). The empty sequence.

singleton :: a -> Seq a

\( O(1) \). A singleton sequence.

(<|) :: a -> Seq a -> Seq a infixr 5

\( O(1) \). Add an element to the left end of a sequence. Mnemonic: a triangle with the single element at the pointy end.

(|>) :: Seq a -> a -> Seq a infixl 5

\( O(1) \). Add an element to the right end of a sequence. Mnemonic: a triangle with the single element at the pointy end.

(><) :: Seq a -> Seq a -> Seq a infixr 5

\( O(\log(\min(n_1,n_2))) \). Concatenate two sequences.

fromList :: [a] -> Seq a

\( O(n) \). Create a sequence from a finite list of elements. There is a function toList in the opposite direction for all instances of the Foldable class, including Seq.

fromFunction :: Int -> (Int -> a) -> Seq a

\( O(n) \). Convert a given sequence length and a function representing that sequence into a sequence.

Since: 0.5.6.2

fromArray :: Ix i => Array i a -> Seq a

\( O(n) \). Create a sequence consisting of the elements of an Array. Note that the resulting sequence elements may be evaluated lazily (as on GHC), so you must force the entire structure to be sure that the original array can be garbage-collected.

Since: 0.5.6.2

Repetition

replicate :: Int -> a -> Seq a

\( O(\log n) \). replicate n x is a sequence consisting of n copies of x.

replicateA :: Applicative f => Int -> f a -> f (Seq a)

replicateA is an Applicative version of replicate, and makes \( O(\log n) \) calls to liftA2 and pure.

replicateA n x = sequenceA (replicate n x)

replicateM :: Applicative m => Int -> m a -> m (Seq a)

replicateM is a sequence counterpart of replicateM.

replicateM n x = sequence (replicate n x)

For base >= 4.8.0 and containers >= 0.5.11, replicateM is a synonym for replicateA.

cycleTaking :: Int -> Seq a -> Seq a

O(log k). cycleTaking k xs forms a sequence of length k by repeatedly concatenating xs with itself. xs may only be empty if k is 0.

cycleTaking k = fromList . take k . cycle . toList

Iterative construction

iterateN :: Int -> (a -> a) -> a -> Seq a

\( O(n) \). Constructs a sequence by repeated application of a function to a seed value.

iterateN n f x = fromList (Prelude.take n (Prelude.iterate f x))

unfoldr :: (b -> Maybe (a, b)) -> b -> Seq a

Builds a sequence from a seed value. Takes time linear in the number of generated elements. WARNING: If the number of generated elements is infinite, this method will not terminate.

unfoldl :: (b -> Maybe (b, a)) -> b -> Seq a

unfoldl f x is equivalent to reverse (unfoldr (fmap swap . f) x).

Deconstruction

Additional functions for deconstructing sequences are available via the Foldable instance of Seq.

Queries

null :: Seq a -> Bool

\( O(1) \). Is this the empty sequence?

length :: Seq a -> Int

\( O(1) \). The number of elements in the sequence.

Views

data ViewL a

View of the left end of a sequence.

Constructors

EmptyL
empty sequence

a :< (Seq a) infixr 5
leftmost element and the rest of the sequence

viewl :: Seq a -> ViewL a

\( O(1) \). Analyse the left end of a sequence.

data ViewR a

View of the right end of a sequence.

Constructors

EmptyR
empty sequence

(Seq a) :> a infixl 5
the sequence minus the rightmost element, and the rightmost element

viewr :: Seq a -> ViewR a

\( O(1) \). Analyse the right end of a sequence.

Scans

scanl :: (a -> b -> a) -> a -> Seq b -> Seq a

scanl is similar to foldl, but returns a sequence of reduced values from the left:

scanl f z (fromList [x1, x2, ...]) = fromList [z, z `f` x1, (z `f` x1) `f` x2, ...]

scanl1 :: (a -> a -> a) -> Seq a -> Seq a

scanl1 is a variant of scanl that has no starting value argument:

scanl1 f (fromList [x1, x2, ...]) = fromList [x1, x1 `f` x2, ...]

scanr :: (a -> b -> b) -> b -> Seq a -> Seq b

scanr is the right-to-left dual of scanl.

scanr1 :: (a -> a -> a) -> Seq a -> Seq a

scanr1 is a variant of scanr that has no starting value argument.

Sublists

tails :: Seq a -> Seq (Seq a)

\( O(n) \). Returns a sequence of all suffixes of this sequence, longest first. For example,

tails (fromList "abc") = fromList [fromList "abc", fromList "bc", fromList "c", fromList ""]

Evaluating the \( i \)th suffix takes \( O(\log(\min(i, n-i))) \), but evaluating every suffix in the sequence takes \( O(n) \) due to sharing.

inits :: Seq a -> Seq (Seq a)

\( O(n) \). Returns a sequence of all prefixes of this sequence, shortest first. For example,

inits (fromList "abc") = fromList [fromList "", fromList "a", fromList "ab", fromList "abc"]

Evaluating the \( i \)th prefix takes \( O(\log(\min(i, n-i))) \), but evaluating every prefix in the sequence takes \( O(n) \) due to sharing.

chunksOf :: Int -> Seq a -> Seq (Seq a)

\(O \Bigl(\bigl(\frac{n}{c}\bigr) \log c\Bigr)\). chunksOf c xs splits xs into chunks of size c>0. If c does not divide the length of xs evenly, then the last element of the result will be short.

Side note: the given performance bound is missing some messy terms that only really affect edge cases. Performance degrades smoothly from \( O(1) \) (for \( c = n \)) to \( O(n) \) (for \( c = 1 \)). The true bound is more like \( O \Bigl( \bigl(\frac{n}{c} - 1\bigr) (\log (c + 1)) + 1 \Bigr) \)

Since: 0.5.8

Sequential searches

takeWhileL :: (a -> Bool) -> Seq a -> Seq a

\( O(i) \) where \( i \) is the prefix length. takeWhileL, applied to a predicate p and a sequence xs, returns the longest prefix (possibly empty) of xs of elements that satisfy p.

takeWhileR :: (a -> Bool) -> Seq a -> Seq a

\( O(i) \) where \( i \) is the suffix length. takeWhileR, applied to a predicate p and a sequence xs, returns the longest suffix (possibly empty) of xs of elements that satisfy p.

takeWhileR p xs is equivalent to reverse (takeWhileL p (reverse xs)).

dropWhileL :: (a -> Bool) -> Seq a -> Seq a

\( O(i) \) where \( i \) is the prefix length. dropWhileL p xs returns the suffix remaining after takeWhileL p xs.

dropWhileR :: (a -> Bool) -> Seq a -> Seq a

\( O(i) \) where \( i \) is the suffix length. dropWhileR p xs returns the prefix remaining after takeWhileR p xs.

dropWhileR p xs is equivalent to reverse (dropWhileL p (reverse xs)).

spanl :: (a -> Bool) -> Seq a -> (Seq a, Seq a)

\( O(i) \) where \( i \) is the prefix length. spanl, applied to a predicate p and a sequence xs, returns a pair whose first element is the longest prefix (possibly empty) of xs of elements that satisfy p and the second element is the remainder of the sequence.

spanr :: (a -> Bool) -> Seq a -> (Seq a, Seq a)

\( O(i) \) where \( i \) is the suffix length. spanr, applied to a predicate p and a sequence xs, returns a pair whose first element is the longest suffix (possibly empty) of xs of elements that satisfy p and the second element is the remainder of the sequence.

breakl :: (a -> Bool) -> Seq a -> (Seq a, Seq a)

\( O(i) \) where \( i \) is the breakpoint index. breakl, applied to a predicate p and a sequence xs, returns a pair whose first element is the longest prefix (possibly empty) of xs of elements that do not satisfy p and the second element is the remainder of the sequence.

breakl p is equivalent to spanl (not . p).

breakr :: (a -> Bool) -> Seq a -> (Seq a, Seq a)

breakr p is equivalent to spanr (not . p).

partition :: (a -> Bool) -> Seq a -> (Seq a, Seq a)

\( O(n) \). The partition function takes a predicate p and a sequence xs and returns sequences of those elements which do and do not satisfy the predicate.

filter :: (a -> Bool) -> Seq a -> Seq a

\( O(n) \). The filter function takes a predicate p and a sequence xs and returns a sequence of those elements which satisfy the predicate.

Sorting

sort :: Ord a => Seq a -> Seq a

\( O(n \log n) \). sort sorts the specified Seq by the natural ordering of its elements. The sort is stable. If stability is not required, unstableSort can be slightly faster.

sortBy :: (a -> a -> Ordering) -> Seq a -> Seq a

\( O(n \log n) \). sortBy sorts the specified Seq according to the specified comparator. The sort is stable. If stability is not required, unstableSortBy can be slightly faster.

sortOn :: Ord b => (a -> b) -> Seq a -> Seq a

\( O(n \log n) \). sortOn sorts the specified Seq by comparing the results of a key function applied to each element. sortOn f is equivalent to sortBy (compare `on` f), but has the performance advantage of only evaluating f once for each element in the input list. This is called the decorate-sort-undecorate paradigm, or Schwartzian transform.

An example of using sortOn might be to sort a Seq of strings according to their length:

sortOn length (fromList ["alligator", "monkey", "zebra"]) == fromList ["zebra", "monkey", "alligator"]

If, instead, sortBy had been used, length would be evaluated on every comparison, giving \( O(n \log n) \) evaluations, rather than \( O(n) \).

If f is very cheap (for example a record selector, or fst), sortBy (compare `on` f) will be faster than sortOn f.

unstableSort :: Ord a => Seq a -> Seq a

\( O(n \log n) \). unstableSort sorts the specified Seq by the natural ordering of its elements, but the sort is not stable. This algorithm is frequently faster and uses less memory than sort.

unstableSortBy :: (a -> a -> Ordering) -> Seq a -> Seq a

\( O(n \log n) \). A generalization of unstableSort, unstableSortBy takes an arbitrary comparator and sorts the specified sequence. The sort is not stable. This algorithm is frequently faster and uses less memory than sortBy.

unstableSortOn :: Ord b => (a -> b) -> Seq a -> Seq a

\( O(n \log n) \). unstableSortOn sorts the specified Seq by comparing the results of a key function applied to each element. unstableSortOn f is equivalent to unstableSortBy (compare `on` f), but has the performance advantage of only evaluating f once for each element in the input list. This is called the decorate-sort-undecorate paradigm, or Schwartzian transform.

An example of using unstableSortOn might be to sort a Seq of strings according to their length:

unstableSortOn length (fromList ["alligator", "monkey", "zebra"]) == fromList ["zebra", "monkey", "alligator"]

If, instead, unstableSortBy had been used, length would be evaluated on every comparison, giving \( O(n \log n) \) evaluations, rather than \( O(n) \).

If f is very cheap (for example a record selector, or fst), unstableSortBy (compare `on` f) will be faster than unstableSortOn f.

Indexing

lookup :: Int -> Seq a -> Maybe a

\( O(\log(\min(i,n-i))) \). The element at the specified position, counting from 0. If the specified position is negative or at least the length of the sequence, lookup returns Nothing.

0 <= i < length xs ==> lookup i xs == Just (toList xs !! i)

i < 0 || i >= length xs ==> lookup i xs = Nothing

Unlike index, this can be used to retrieve an element without forcing it. For example, to insert the fifth element of a sequence xs into a Map m at key k, you could use

case lookup 5 xs of
  Nothing -> m
  Just x -> insert k x m

Since: 0.5.8

(!?) :: Seq a -> Int -> Maybe a

\( O(\log(\min(i,n-i))) \). A flipped, infix version of lookup.

Since: 0.5.8

index :: Seq a -> Int -> a

\( O(\log(\min(i,n-i))) \). The element at the specified position, counting from 0. The argument should thus be a non-negative integer less than the size of the sequence. If the position is out of range, index fails with an error.

xs `index` i = toList xs !! i

Caution: index necessarily delays retrieving the requested element until the result is forced. It can therefore lead to a space leak if the result is stored, unforced, in another structure. To retrieve an element immediately without forcing it, use lookup or '(!?)'.

adjust :: (a -> a) -> Int -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). Update the element at the specified position. If the position is out of range, the original sequence is returned. adjust can lead to poor performance and even memory leaks, because it does not force the new value before installing it in the sequence. adjust' should usually be preferred.

Since: 0.5.8

adjust' :: forall a. (a -> a) -> Int -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). Update the element at the specified position. If the position is out of range, the original sequence is returned. The new value is forced before it is installed in the sequence.

adjust' f i xs =
 case xs !? i of
   Nothing -> xs
   Just x -> let !x' = f x
             in update i x' xs

Since: 0.5.8

update :: Int -> a -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). Replace the element at the specified position. If the position is out of range, the original sequence is returned.

take :: Int -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). The first i elements of a sequence. If i is negative, take i s yields the empty sequence. If the sequence contains fewer than i elements, the whole sequence is returned.

drop :: Int -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). Elements of a sequence after the first i. If i is negative, drop i s yields the whole sequence. If the sequence contains fewer than i elements, the empty sequence is returned.

insertAt :: Int -> a -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). insertAt i x xs inserts x into xs at the index i, shifting the rest of the sequence over.

insertAt 2 x (fromList [a,b,c,d]) = fromList [a,b,x,c,d]
insertAt 4 x (fromList [a,b,c,d]) = insertAt 10 x (fromList [a,b,c,d])
                                  = fromList [a,b,c,d,x]

insertAt i x xs = take i xs >< singleton x >< drop i xs

Since: 0.5.8

deleteAt :: Int -> Seq a -> Seq a

\( O(\log(\min(i,n-i))) \). Delete the element of a sequence at a given index. Return the original sequence if the index is out of range.

deleteAt 2 [a,b,c,d] = [a,b,d]
deleteAt 4 [a,b,c,d] = deleteAt (-1) [a,b,c,d] = [a,b,c,d]

Since: 0.5.8

splitAt :: Int -> Seq a -> (Seq a, Seq a)

\( O(\log(\min(i,n-i))) \). Split a sequence at a given position. splitAt i s = (take i s, drop i s).

Indexing with predicates

These functions perform sequential searches from the left or right ends of the sequence, returning indices of matching elements.

elemIndexL :: Eq a => a -> Seq a -> Maybe Int

elemIndexL finds the leftmost index of the specified element, if it is present, and otherwise Nothing.

elemIndicesL :: Eq a => a -> Seq a -> [Int]

elemIndicesL finds the indices of the specified element, from left to right (i.e. in ascending order).

elemIndexR :: Eq a => a -> Seq a -> Maybe Int

elemIndexR finds the rightmost index of the specified element, if it is present, and otherwise Nothing.

elemIndicesR :: Eq a => a -> Seq a -> [Int]

elemIndicesR finds the indices of the specified element, from right to left (i.e. in descending order).

findIndexL :: (a -> Bool) -> Seq a -> Maybe Int

findIndexL p xs finds the index of the leftmost element that satisfies p, if any exist.

findIndicesL :: (a -> Bool) -> Seq a -> [Int]

findIndicesL p finds all indices of elements that satisfy p, in ascending order.

findIndexR :: (a -> Bool) -> Seq a -> Maybe Int

findIndexR p xs finds the index of the rightmost element that satisfies p, if any exist.

findIndicesR :: (a -> Bool) -> Seq a -> [Int]

findIndicesR p finds all indices of elements that satisfy p, in descending order.

Folds

General folds are available via the Foldable instance of Seq.

foldMapWithIndex :: Monoid m => (Int -> a -> m) -> Seq a -> m

foldlWithIndex :: (b -> Int -> a -> b) -> b -> Seq a -> b

foldlWithIndex is a version of foldl that also provides access to the index of each element.

foldrWithIndex :: (Int -> a -> b -> b) -> b -> Seq a -> b

foldrWithIndex is a version of foldr that also provides access to the index of each element.

Transformations

mapWithIndex :: (Int -> a -> b) -> Seq a -> Seq b

A generalization of fmap, mapWithIndex takes a mapping function that also depends on the element's index, and applies it to every element in the sequence.

traverseWithIndex :: Applicative f => (Int -> a -> f b) -> Seq a -> f (Seq b)

traverseWithIndex is a version of traverse that also offers access to the index of each element.

Since: 0.5.8

reverse :: Seq a -> Seq a

\( O(n) \). The reverse of a sequence.

intersperse :: a -> Seq a -> Seq a

\( O(n) \). Intersperse an element between the elements of a sequence.

intersperse a empty = empty
intersperse a (singleton x) = singleton x
intersperse a (fromList [x,y]) = fromList [x,a,y]
intersperse a (fromList [x,y,z]) = fromList [x,a,y,a,z]

Since: 0.5.8

Zips and unzip

zip :: Seq a -> Seq b -> Seq (a, b)

\( O(\min(n_1,n_2)) \). zip takes two sequences and returns a sequence of corresponding pairs. If one input is short, excess elements are discarded from the right end of the longer sequence.

zipWith :: (a -> b -> c) -> Seq a -> Seq b -> Seq c

\( O(\min(n_1,n_2)) \). zipWith generalizes zip by zipping with the function given as the first argument, instead of a tupling function. For example, zipWith (+) is applied to two sequences to take the sequence of corresponding sums.

zip3 :: Seq a -> Seq b -> Seq c -> Seq (a, b, c)

\( O(\min(n_1,n_2,n_3)) \). zip3 takes three sequences and returns a sequence of triples, analogous to zip.

zipWith3 :: (a -> b -> c -> d) -> Seq a -> Seq b -> Seq c -> Seq d

\( O(\min(n_1,n_2,n_3)) \). zipWith3 takes a function which combines three elements, as well as three sequences and returns a sequence of their point-wise combinations, analogous to zipWith.

zip4 :: Seq a -> Seq b -> Seq c -> Seq d -> Seq (a, b, c, d)

\( O(\min(n_1,n_2,n_3,n_4)) \). zip4 takes four sequences and returns a sequence of quadruples, analogous to zip.

zipWith4 :: (a -> b -> c -> d -> e) -> Seq a -> Seq b -> Seq c -> Seq d -> Seq e

\( O(\min(n_1,n_2,n_3,n_4)) \). zipWith4 takes a function which combines four elements, as well as four sequences and returns a sequence of their point-wise combinations, analogous to zipWith.

unzip :: Seq (a, b) -> (Seq a, Seq b)

Unzip a sequence of pairs.

unzip ps = ps `seq' (fmap fst ps) (fmap snd ps)

Example:

unzip $ fromList [(1,"a"), (2,"b"), (3,"c")] =
  (fromList [1,2,3], fromList ["a", "b", "c"])

See the note about efficiency at unzipWith.

Since: 0.5.11

unzipWith :: (a -> (b, c)) -> Seq a -> (Seq b, Seq c)

\( O(n) \). Unzip a sequence using a function to divide elements.

 unzipWith f xs == unzip (fmap f xs)

Efficiency note:

unzipWith produces its two results in lockstep. If you calculate unzipWith f xs and fully force either of the results, then the entire structure of the other one will be built as well. This behavior allows the garbage collector to collect each calculated pair component as soon as it dies, without having to wait for its mate to die. If you do not need this behavior, you may be better off simply calculating the sequence of pairs and using fmap to extract each component sequence.

Since: 0.5.11